fixed fatal error

+++ src/main/java/epicentr/repositories/UserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Ramesh Fadatare
  *
  */
-public interface UserRepository extends JpaRepository<User, Long>
+public interface UserRepository extends JpaRepository<User, Integer>
 {
 
 	User findByEmail(String email);
